refac: moving blockchain_client.rs to importer's module
