+++ src/eth/follower/importer/importer_config.rs
@@ -10,11 +10,11 @@ use crate::NodeMode;
 use crate::eth::executor::Executor;
 use crate::eth::follower::ConsensusError;
 use crate::eth::follower::ImporterError;
+use crate::eth::follower::importer::BlockchainClient;
 use crate::eth::follower::importer::ImporterMode;
 use crate::eth::follower::importer::importer_supervisor::ImporterConsensus;
 use crate::eth::follower::importer::importer_supervisor::start_importer;
 use crate::eth::miner::Miner;
-use crate::eth::rpc::BlockchainClient;
 use crate::eth::rpc::RpcContext;
 use crate::eth::rpc::pagination::validate_response_size_limit;
 use crate::eth::storage::StratusStorage;
