+++ src/eth/follower/importer/importer_supervisor.rs
@@ -10,6 +10,7 @@ use crate::eth::executor::Executor;
 use crate::eth::follower::consensus::Consensus;
 use crate::eth::follower::consensus::LagDirection;
 use crate::eth::follower::consensus::LagStatus;
+use crate::eth::follower::importer::BlockchainClient;
 use crate::eth::follower::importer::EXTERNAL_RPC_CURRENT_BLOCK;
 use crate::eth::follower::importer::ImporterMode;
 use crate::eth::follower::importer::LATEST_FETCHED_BLOCK_TIME;
@@ -23,7 +24,6 @@ use crate::eth::follower::importer::importers::fake_leader::FakeLeaderWorker;
 use crate::eth::follower::importer::importers::replication::ReplicationWorker;
 use crate::eth::follower::importer::start_number_fetcher;
 use crate::eth::miner::Miner;
-use crate::eth::rpc::BlockchainClient;
 use crate::eth::storage::StratusStorage;
 use crate::eth::types::BlockNumber;
 use crate::ext::spawn;
