--- src/eth/rpc/blockchain_client/mod.rs
@@ -1,4 +1,0 @@
-#[allow(clippy::module_inception)]
-pub mod blockchain_client;
-
-pub use blockchain_client::BlockchainClient;
